mudei a lógica por trás do delete de usuário e filme, e por fim, adicionei um teste para a rota de delete do aluguel

+++ src/modules/movies/deleteMovies/DeleteMovieUseCase.ts
@@ -9,12 +9,25 @@ interface CreateMovieResult {
 export async function DeleteMovieUseCase({
   id,
 }: any): Promise<CreateMovieResult> {
-  const MovieDelete = await prisma.movie.delete({
+  const movieExists = await prisma.movie.findUnique({
     where: {
       id,
     },
   });
-  
+
+  if (!movieExists) {
+    return {
+      success: false,
+      error: "Não foi possível excluir o filme do sistema, pois ele não foi encontrado",
+    };
+  }
+
+  await prisma.movie.delete({
+    where: {
+      id,
+    },
+  });
+
   return {
     success: true,
     data: "Produto Excluído com sucesso",
